Renamed main.py, able to import as module now.
